+++ src/main/java/code/webdkmh/utli/security/MySimpleUrlAuthenticationSuccessHandler.java
@@ -1,4 +1,4 @@
-package code.webdkmh.security;
+package code.webdkmh.utli.security;
 
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.GrantedAuthority;
